Next button not handled correctly
